goodix_driver_gt9886: use u8 instead of int for udfps* and double_tap*
Save them bytes

+++ drivers/input/touchscreen/xiaomi/goodix_driver_gt9886/goodix_ts_core.h
@@ -478,10 +478,10 @@ struct goodix_ts_core {
 	int aod_status;
 	int fod_pressed;
 	int fod_test;
-	int udfps_pressed;
-	int udfps_enabled;
-	int double_tap_pressed;
-	int double_tap_enabled;
+	u8 udfps_pressed;
+	u8 udfps_enabled;
+	u8 double_tap_pressed;
+	u8 double_tap_enabled;
 	int result_type;
 	struct class *gtp_tp_class;
 	struct device *gtp_touch_dev;
